requirements.txt
Django==4.1
djangorestframework==3.13.1
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-extensions==3.2.0
Pillow==9.2.0

pip install -r requirements.txt
